Struct plonk_15_wires_circuits::gate::CircuitGate

pub struct CircuitGate<F: FftField> {
    pub row: usize,
    pub typ: GateType,
    pub wires: GateWires,
    pub c: Vec<F>,
}

Fields
row: usize

row position in the circuit

typ: GateType

type of the gate

wires: GateWires

gate wires

c: Vec<F>

constraints vector

Implementations

impl<F: FftField> CircuitGate<F>

pub fn zero(row: usize, wires: GateWires) -> Self

this function creates “empty” circuit gate

pub fn verify(
&self,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

This function verifies the consistency of the wire assignements (witness) against the constraints

impl<F: FftField> CircuitGate<F>

pub fn create_add(row: usize, wires: GateWires) -> Self

Creates an ECC add gate. Warning: this assumes that the two points are:

on the curve

not mutual inverses (e.g. P and -P)

not the same point (e.g. P and P)

pub fn verify_add(&self, witness: &[Vec<F>; 15]) -> Result<(), String>

Given a set of [witness] over an Add gate, verify that the constraints checks out.

pub fn add(&self) -> F

Returns 1 if self is an GateType::Add gate, 0 otherwise.

impl<F: FftField> CircuitGate<F>

pub fn create_double(row: usize, wires: GateWires) -> Self

pub fn verify_double(&self, witness: &[Vec<F>; 15]) -> Result<(), String>

pub fn double(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_endomul(row: usize, wires: GateWires) -> Self

pub fn verify_endomul(
&self,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

pub fn endomul(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_generic(
row: usize,
wires: GateWires,
qw: [F; 3],
qm: F,
qc: F
) -> Self

pub fn create_generic_add(row: usize, wires: GateWires) -> Self

creates an addition gate

pub fn create_generic_mul(row: usize, wires: GateWires) -> Self

creates a multiplication gate

pub fn create_generic_const(row: usize, wires: GateWires, constant: F) -> Self

creates a constant gate

pub fn create_generic_public(row: usize, wires: GateWires) -> Self

creates a public input gate

pub fn verify_generic(&self, witness: &[Vec<F>; 15]) -> Result<(), String>

verifies that the generic gate constraints are solved by the witness

pub fn generic(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_lookup(row: usize, wires: GateWires) -> Self

pub fn verify_lookup(&self, witness: &[Vec<F>; 15]) -> Result<(), String>

pub fn lookup(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_poseidon(row: usize, wires: GateWires, c: [[F; 3]; 5]) -> Self

pub fn create_poseidon_gadget(
row: usize,
first_and_last_row: [GateWires; 2],
round_constants: &Vec<Vec<F>>
) -> (Vec<Self>, usize)

create_poseidon_gadget(row, first_and_last_row, round_constants) creates an entire set of constraint for a Poseidon hash. For that, you need to pass:

the index of the first row

the first and last rows’ wires (because they are used in the permutation)

the round constants The function returns a set of gates, as well as the next pointer to the circuit (next empty absolute row)

pub fn verify_poseidon(
&self,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

Checks if a witness verifies a poseidon gate

pub fn ps(&self) -> F

pub fn rc(&self) -> [[F; 3]; 5]

round constant that are relevant for this specific gate

impl<F: FftField> CircuitGate<F>

pub fn create_vbmul(row: usize, wires: &[GateWires; 3]) -> Vec<Self>

pub fn verify_vbmul(&self, witness: &[Vec<F>; 15]) -> Result<(), String>

pub fn vbmul(&self) -> F