Struct plonk_15_wires_circuits::nolookup::constraints::ConstraintSystem

pub struct ConstraintSystem<F: FftField> {

    pub public: usize,
    pub domain: EvaluationDomains<F>,
    pub gates: Vec<CircuitGate<F>>,
    pub sigmam: [DP<F>; 7],
    pub zkpm: DP<F>,
    pub qwm: [DP<F>; 3],
    pub qmm: DP<F>,
    pub qc: DP<F>,
    pub rcm: [[DP<F>; 3]; 5],
    pub psm: DP<F>,
    pub addm: DP<F>,
    pub doublem: DP<F>,
    pub mulm: DP<F>,
    pub emulm: DP<F>,
    pub qwl: [E<F, D<F>>; 3],
    pub qml: E<F, D<F>>,
    pub sigmal1: [Vec<F>; 7],
    pub sigmal8: [E<F, D<F>>; 7],
    pub sid: Vec<F>,
    pub ps4: E<F, D<F>>,
    pub ps8: E<F, D<F>>,
    pub addl: E<F, D<F>>,
    pub doubl8: E<F, D<F>>,
    pub doubl4: E<F, D<F>>,
    pub mull4: E<F, D<F>>,
    pub mull8: E<F, D<F>>,
    pub emull: E<F, D<F>>,
    pub l1: E<F, D<F>>,
    pub l04: E<F, D<F>>,
    pub l08: E<F, D<F>>,
    pub zero4: E<F, D<F>>,
    pub zero8: E<F, D<F>>,
    pub zkpl: E<F, D<F>>,
    pub shift: [F; 7],
    pub endo: F,
    pub fr_sponge_params: ArithmeticSpongeParams<F>,

}

Fields
public: usize

number of public inputs

domain: EvaluationDomains<F>

evaluation domains

gates: Vec<CircuitGate<F>>

circuit gates

sigmam: [DP<F>; 7]

permutation polynomial array

zkpm: DP<F>

zero-knowledge polynomial

qwm: [DP<F>; 3]

linear wire constraint polynomial

qmm: DP<F>

multiplication polynomial

qc: DP<F>

constant wire polynomial

rcm: [[DP<F>; 3]; 5]

round constant polynomials

psm: DP<F>

poseidon constraint selector polynomial

addm: DP<F>

EC point addition constraint selector polynomial

doublem: DP<F>

EC point doubling constraint selector polynomial

mulm: DP<F>

mulm constraint selector polynomial

emulm: DP<F>

emulm constraint selector polynomial

qwl: [E<F, D<F>>; 3]

left input wire polynomial over domain.d4

qml: E<F, D<F>>

multiplication evaluations over domain.d4

sigmal1: [Vec<F>; 7]

permutation polynomial array evaluations over domain d1

sigmal8: [E<F, D<F>>; 7]

permutation polynomial array evaluations over domain d8

sid: Vec<F>

SID polynomial

ps4: E<F, D<F>>

poseidon selector over domain.d4

ps8: E<F, D<F>>

poseidon selector over domain.d8

addl: E<F, D<F>>

EC point addition selector evaluations w over domain.d4

doubl8: E<F, D<F>>

EC point doubling selector evaluations w over domain.d8

doubl4: E<F, D<F>>

EC point doubling selector evaluations w over domain.d4

mull4: E<F, D<F>>

scalar multiplication selector evaluations over domain.d4

mull8: E<F, D<F>>

scalar multiplication selector evaluations over domain.d8

emull: E<F, D<F>>

endoscalar multiplication selector evaluations over domain.d8

l1: E<F, D<F>>

1-st Lagrange evaluated over domain.d8

l04: E<F, D<F>>

0-th Lagrange evaluated over domain.d4

l08: E<F, D<F>>

0-th Lagrange evaluated over domain.d8

zero4: E<F, D<F>>

zero evaluated over domain.d8

zero8: E<F, D<F>>

zero evaluated over domain.d8

zkpl: E<F, D<F>>

zero-knowledge polynomial over domain.d8

shift: [F; 7]

wire coordinate shifts

endo: F

coefficient for the group endomorphism

fr_sponge_params: ArithmeticSpongeParams<F>

random oracle argument parameters

Implementations

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn create(
gates: Vec<CircuitGate<F>>,
fr_sponge_params: ArithmeticSpongeParams<F>,
public: usize
) -> Option<Self>

creates a constraint system from a vector of gates (CircuitGate), some sponge parameters (ArithmeticSpongeParams), and the number of public inputs.

pub fn verify(&self, witness: &[Vec<F>; 15]) -> Result<(), GateError>

This function verifies the consistency of the wire assignements (witness) against the constraints witness: wire assignement witness RETURN: verification status

pub fn sample_shift(domain: &D<F>, i: &mut u32) -> F

sample coordinate shifts deterministically

pub fn sample_shifts(domain: &D<F>, len: usize) -> Vec<F>

pub fn evaluate(&self, w: &[DP<F>; 15], z: &DP<F>) -> WitnessOverDomains<F>

evaluate witness polynomials over domains

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn ecad_quot(
&self,
polys: &WitnessOverDomains<F>,
alpha: &[F]
) -> Evaluations<F, D<F>>

pub fn ecad_scalars(evals: &Vec<ProofEvaluations<F>>, alpha: &[F]) -> F

pub fn ecad_lnrz(
&self,
evals: &Vec<ProofEvaluations<F>>,
alpha: &[F]
) -> DensePolynomial<F>

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn double_quot(
&self,
polys: &WitnessOverDomains<F>,
alpha: &[F]
) -> (Evaluations<F, D<F>>, Evaluations<F, D<F>>)

pub fn double_scalars(evals: &Vec<ProofEvaluations<F>>, alpha: &[F]) -> F

pub fn double_lnrz(
&self,
evals: &Vec<ProofEvaluations<F>>,
alpha: &[F]
) -> DensePolynomial<F>

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn endomul_quot(
&self,
polys: &WitnessOverDomains<F>,
alpha: &[F]
) -> Evaluations<F, D<F>>

pub fn endomul_scalars(
evals: &Vec<ProofEvaluations<F>>,
endo: F,
alpha: &[F]
) -> F

pub fn endomul_lnrz(
&self,
evals: &Vec<ProofEvaluations<F>>,
alpha: &[F]
) -> DensePolynomial<F>

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn gnrc_quot(
&self,
witness_d4: &[Evaluations<F, D<F>>; 15],
public: &DensePolynomial<F>
) -> (Evaluations<F, D<F>>, DensePolynomial<F>)

generic constraint quotient poly contribution computation

pub fn gnrc_scalars(w_zeta: &[F; 15]) -> Vec<F>

produces w0 * w1, w0, w1, w2, 1

pub fn gnrc_lnrz(&self, w_zeta: &[F; 15]) -> DensePolynomial<F>

generic constraint linearization poly contribution computation

pub fn verify_generic(
&self,
witness: &[DensePolynomial<F>; 15],
public: &DensePolynomial<F>
) -> bool

function to verify the generic polynomials with a witness

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn perm_quot(
&self,
lagrange: &WitnessOverDomains<F>,
beta: F,
gamma: F,
z: &DensePolynomial<F>,
alpha: &[F]
) -> Result<(Evaluations<F, D<F>>, DensePolynomial<F>), ProofError>

permutation quotient poly contribution computation

pub fn perm_lnrz(
&self,
e: &Vec<ProofEvaluations<F>>,
zeta: F,
beta: F,
gamma: F,
alpha: &[F]
) -> DensePolynomial<F>

permutation linearization poly contribution computation

pub fn perm_scalars(
e: &Vec<ProofEvaluations<F>>,
beta: F,
gamma: F,
alpha: &[F],
zkp_zeta: F
) -> F

pub fn perm_aggreg(
&self,
witness: &[Vec<F>; 15],
beta: &F,
gamma: &F,
rng: &mut ThreadRng
) -> Result<DensePolynomial<F>, ProofError>

permutation aggregation polynomial computation

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn psdn_quot(
&self,
polys: &WitnessOverDomains<F>,
params: &ArithmeticSpongeParams<F>,
alpha: &[F]
) -> (Evaluations<F, D<F>>, Evaluations<F, D<F>>, DensePolynomial<F>)

poseidon quotient poly contribution computation f^7 + c(x) - f(wx) optimization: shuffle the intra-row rounds so that the final state is in one of the permutation columns

pub fn psdn_scalars(
evals: &Vec<ProofEvaluations<F>>,
params: &ArithmeticSpongeParams<F>,
alpha: &[F]
) -> Vec<F>

pub fn psdn_lnrz(
&self,
evals: &Vec<ProofEvaluations<F>>,
params: &ArithmeticSpongeParams<F>,
alpha: &[F]
) -> DensePolynomial<F>

poseidon linearization poly contribution computation f^7 + c(x) - f(wx)

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn vbmul_quot(
&self,
polys: &WitnessOverDomains<F>,
alpha: &[F]
) -> (Evaluations<F, D<F>>, Evaluations<F, D<F>>)

pub fn vbmul_scalars(evals: &Vec<ProofEvaluations<F>>, alpha: &[F]) -> F

pub fn vbmul_lnrz(
&self,
evals: &Vec<ProofEvaluations<F>>,
alpha: &[F]
) -> DensePolynomial<F>